requirements.txt
requests==2.9.1
flake8==2.5.4
nose==1.3.7
mock==1.3.0
splinter==0.7.3
pyfunctional==0.6.0
radon==1.4.0
